Move trimZeroes to the single using class.
